+++ example/android/app/src/main/java/com/airbnb/android/react/maps/example/MainActivity.java
@@ -4,14 +4,12 @@ import com.facebook.react.ReactActivity;
 
 
 public class MainActivity extends ReactActivity {
-
-    /**
-     * Returns the name of the main component registered from JavaScript.
-     * This is used to schedule rendering of the component.
-     */
-    @Override
-    protected String getMainComponentName() {
-        return "AirMapsExplorer";
-    }
-
+  /**
+   * Returns the name of the main component registered from JavaScript.
+   * This is used to schedule rendering of the component.
+   */
+  @Override
+  protected String getMainComponentName() {
+    return "AirMapsExplorer";
+  }
 }