+++ example/examples/CustomOverlayXMarksTheSpot.js
@@ -1,4 +1,6 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
+
 import { View } from 'react-native';
 import MapView from 'react-native-maps';