+++ lib/ios/AirGoogleMaps/AIRGoogleMap.m
@@ -155,6 +155,10 @@ id regionAsJSON(MKCoordinateRegion region) {
   self.camera = [AIRGoogleMap makeGMSCameraPositionFromMap:self  andMKCoordinateRegion:region];
 }
 
+- (void)didPrepareMap {
+    if (self.onMapReady) self.onMapReady(@{});
+}
+
 - (BOOL)didTapMarker:(GMSMarker *)marker {
   AIRGMSMarker *airMarker = (AIRGMSMarker *)marker;
 
@@ -168,6 +172,16 @@ id regionAsJSON(MKCoordinateRegion region) {
   // TODO: not sure why this is necessary
   [self setSelectedMarker:marker];
   return NO;
+}
+
+- (void)didTapPolyline:(GMSOverlay *)polyline {
+  AIRGMSPolyline *airPolyline = (AIRGMSPolyline *)polyline;
+
+  id event = @{@"action": @"polyline-press",
+               @"id": airPolyline.identifier ?: @"unknown",
+               };
+
+   if (airPolyline.onPress) airPolyline.onPress(event);
 }
 
 - (void)didTapPolygon:(GMSOverlay *)polygon {
@@ -205,6 +219,13 @@ id regionAsJSON(MKCoordinateRegion region) {
   if (self.onChange) self.onChange(event);  // complete
 }
 
+- (void)setMapPadding:(UIEdgeInsets)mapPadding {
+  self.padding = mapPadding;
+}
+
+- (UIEdgeInsets)mapPadding {
+  return self.padding;
+}
 
 - (void)setScrollEnabled:(BOOL)scrollEnabled {
   self.settings.scrollGestures = scrollEnabled;
@@ -306,6 +327,21 @@ id regionAsJSON(MKCoordinateRegion region) {
   return self.settings.myLocationButton;
 }
 
+- (void)setMinZoomLevel:(CGFloat)minZoomLevel {
+  [self setMinZoom:minZoomLevel maxZoom:self.maxZoom ];
+}
+
+- (void)setMaxZoomLevel:(CGFloat)maxZoomLevel {
+  [self setMinZoom:self.minZoom maxZoom:maxZoomLevel ];
+}
+
+- (void)setShowsIndoorLevelPicker:(BOOL)showsIndoorLevelPicker {
+  self.settings.indoorPicker = showsIndoorLevelPicker;
+}
+
+- (BOOL)showsIndoorLevelPicker {
+  return self.settings.indoorPicker;
+}
 
 + (MKCoordinateRegion) makeGMSCameraPositionFromMap:(GMSMapView *)map andGMSCameraPosition:(GMSCameraPosition *)position {
   // solution from here: http://stackoverflow.com/a/16587735/1102215