+++ lib/ios/AirGoogleMaps/AIRGoogleMapMarker.m
@@ -36,6 +36,7 @@ CGRect unionRect(CGRect a, CGRect b) {
   if ((self = [super init])) {
     _realMarker = [[AIRGMSMarker alloc] init];
     _realMarker.fakeMarker = self;
+    _realMarker.tracksViewChanges = true;
   }
   return self;
 }
@@ -155,6 +156,14 @@ CGRect unionRect(CGRect a, CGRect b) {
 
 - (CLLocationCoordinate2D)coordinate {
   return _realMarker.position;
+}
+
+- (void)setRotation:(CLLocationDegrees)rotation {
+    _realMarker.rotation = rotation;
+}
+
+- (CLLocationDegrees)rotation {
+    return _realMarker.rotation;
 }
 
 - (void)setIdentifier:(NSString *)identifier {
@@ -285,6 +294,14 @@ CGRect unionRect(CGRect a, CGRect b) {
 
 - (BOOL)draggable {
   return _realMarker.draggable;
+}
+
+- (void)setTracksViewChanges:(BOOL)tracksViewChanges {
+  _realMarker.tracksViewChanges = tracksViewChanges;
+}
+
+- (BOOL)tracksViewChanges {
+  return _realMarker.tracksViewChanges;
 }
 
 @end